requirements.txt
tensorflow==1.5.0rc1
h5py==2.7.1
pillow==5.0.0
keras==2.1.3
